plugin structure. shortcodes, magic links
