+++ shortcodes/24hour.php
@@ -13,11 +13,11 @@ function dt_24hour_campaign_register_scripts( $atts ){
         wp_localize_script( 'dt_campaign_core', 'dt_campaign_core', [ 'color' => $atts["color"] ?? '' ] );
     }
 
-    wp_enqueue_style( 'dt_campaign_style', trailingslashit( plugin_dir_url( __FILE__ ) ) . '24hour.css', [], filemtime( plugin_dir_path( __FILE__ ) . '24hour.css' ) );
+    wp_enqueue_style( 'dt_campaign_style', trailingslashit( plugin_dir_url( __DIR__ ) ) . 'magic-links/24hour/24hour.css', [], filemtime( plugin_dir_path( __DIR__ ) . 'magic-links/24hour/24hour.css' ) );
 
     //24 hour campaign js
     if ( !wp_script_is( 'dt_campaign' ) ){
-        wp_enqueue_script( 'dt_campaign', trailingslashit( plugin_dir_url( __FILE__ ) ) . '24hour.js', [ 'luxon', 'jquery', 'dt_campaign_core' ], filemtime( plugin_dir_path( __FILE__ ) . '24hour.js' ), true );
+        wp_enqueue_script( 'dt_campaign', trailingslashit( plugin_dir_url( __DIR__ ) ) . 'magic-links/24hour/24hour.js', [ 'luxon', 'jquery', 'dt_campaign_core' ], filemtime( plugin_dir_path( __DIR__ ) . 'magic-links/24hour/24hour.js' ), true );
         wp_localize_script(
             'dt_campaign', 'campaign_objects', [
                 'translations' => [
